+++ colossalai/zero/sharded_model/_utils.py
@@ -1,5 +1,4 @@
-from collections import OrderedDict
-from typing import Any, Callable, Dict, List, Tuple, Union
+from typing import Any, Callable, List, Tuple
 
 import torch
 import torch.nn.functional as F
@@ -10,23 +9,6 @@ def get_gradient_predivide_factor(world_size: int) -> float:
     while world_size % factor == 0 and world_size / factor > factor:
         factor *= 2
     return float(factor)
-
-
-def get_shard(tensor: torch.Tensor, rank: int, world_size: int) -> Tuple[torch.Tensor, int]:
-    """Return the local shard of a full tensor."""
-    # Shard using torch.chunk to match all-gather/reduce-scatter.
-    chunks = list(torch.flatten(tensor).chunk(world_size))
-    while len(chunks) < world_size:
-        chunks.append(chunks[0].new_empty(0))
-
-    # Determine number of padding elements.
-    num_to_pad = chunks[0].numel() - chunks[rank].numel()
-    assert num_to_pad >= 0, num_to_pad
-
-    shard = chunks[rank].clone()
-    if num_to_pad > 0:
-        shard = F.pad(shard, [0, num_to_pad])
-    return shard, num_to_pad
 
 
 def free_storage(data: torch.Tensor) -> None:
@@ -86,31 +68,3 @@ def chunk_and_pad(tensor: torch.Tensor, num_chunks: int) -> List[torch.Tensor]:
     if len(chunks) < num_chunks:
         chunks.extend([torch.zeros_like(chunks[0]) for _ in range(num_chunks - len(chunks))])
     return chunks
-
-
-def assert_in_engine(cond: Any, s: Any) -> None:
-    """Used in backward context to make sure error is printed."""
-    if not cond:
-        print(s)
-        raise AssertionError
-
-
-def replace_state_dict_prefix(state_dict: Union[Dict[str, torch.Tensor], "OrderedDict[str, torch.Tensor]"],
-                              old_prefix: str, new_prefix: str) -> None:
-    """
-    Replace all keys that match a given old_prefix with a new_prefix (in-place).
-
-    Usage::
-
-        state_dict = {"layer.xyz": torch.tensor(1)}
-        replace_state_dict_prefix(state_dict, "layer.", "module.layer.")
-        assert state_dict == {"module.layer.xyz": torch.tensor(1)}
-    """
-    if old_prefix == new_prefix:
-        raise ValueError("old_prefix and new_prefix must be distinct")
-    for key in list(state_dict.keys()):
-        if not key.startswith(old_prefix):
-            continue
-        new_key = new_prefix + key[len(old_prefix):]
-        state_dict[new_key] = state_dict[key]
-        del state_dict[key]
